finished creating a registrer  function
have created functions for a user to registrer an account on the page
