+++ index.js
@@ -1,4 +1,4 @@
-import { apiCall } from "./js/api/api.js";
+import { registerUser } from "./js/entry/register.js";
 
 const logIn = document.getElementById("log-in");
 const register = document.getElementById("register");
@@ -8,19 +8,3 @@ register.addEventListener("submit", (event) => {
   const [userName, email, password] = event.target.elements;
   registerUser(userName, email, password);
 });
-
-const registerUser = (userName, email, password) => {
-  const data = JSON.stringify({
-    name: userName.value,
-    email: email.value,
-    password: password.value,
-  });
-  console.log(data);
-};
-
-const testApi = async () => {
-  const result = await apiCall("jokes/1", "get");
-  console.log(result);
-};
-
-testApi();
